refactor: split generated order controller test (FPLA-1468) (#1041)
* added new tests

* removed test

* remove unused import

* resolve merge conflicts
