+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/GeneratedOrderControllerAboutToStartTest.java
@@ -1,0 +1,58 @@
+package uk.gov.hmcts.reform.fpl.controllers;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.OverrideAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.context.ActiveProfiles;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.fpl.service.time.Time;
+
+import java.util.Map;
+
+import static java.lang.Long.parseLong;
+import static org.assertj.core.api.Assertions.assertThat;
+
+@ActiveProfiles("integration-test")
+@WebMvcTest(GeneratedOrderController.class)
+@OverrideAutoConfiguration(enabled = true)
+public class GeneratedOrderControllerAboutToStartTest extends AbstractControllerTest {
+
+    private static final String CASE_ID = "12345";
+    private final byte[] pdf = {1, 2, 3, 4, 5};
+
+    @Autowired
+    private Time time;
+
+    GeneratedOrderControllerAboutToStartTest() {
+        super("create-order");
+    }
+
+    @Test
+    void shouldReturnErrorsWhenFamilymanNumberIsNotProvided() {
+        CaseDetails caseDetails = CaseDetails.builder()
+            .id(parseLong(CASE_ID))
+            .data(Map.of("data", "some data"))
+            .build();
+
+        AboutToStartOrSubmitCallbackResponse callbackResponse = postAboutToStartEvent(caseDetails);
+
+        assertThat(callbackResponse.getErrors()).containsExactly("Enter Familyman case number");
+    }
+
+    @Test
+    void aboutToStartShouldSetDateOfIssueAsTodayByDefault() {
+        CaseDetails caseDetails = CaseDetails.builder()
+            .id(parseLong(CASE_ID))
+            .data(Map.of("familyManCaseNumber", "123"))
+            .build();
+
+        AboutToStartOrSubmitCallbackResponse callbackResponse = postAboutToStartEvent(caseDetails);
+
+        assertThat(callbackResponse.getErrors()).isEmpty();
+        assertThat(callbackResponse.getData().get("dateOfIssue")).isEqualTo(time.now().toLocalDate().toString());
+    }
+
+
+}
